requirements.txt
# Core Framework
streamlit>=1.28.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0

# PDF Processing
pypdf>=3.17.0
pdf2image>=1.16.3
pillow>=10.0.0

# Vector Database
chromadb>=0.4.15

# HTTP Requests
requests>=2.31.0

# Configuration Management
python-dotenv>=1.0.0

# Standard Library (included in Python, listed for reference)
# os, shutil, tempfile, base64, pathlib, typing

